+++ raycaster_lib/src/render/parallel_renderer/messages.rs
@@ -3,13 +3,13 @@ use super::composition::SubCanvas;
 // Todo combine Vec and pixelbox to newtype
 
 pub struct RenderTask {
-    pub block_id: usize,
-    pub tile_id: usize,
+    pub block_id: u32,
+    pub tile_id: u32,
     pub subcanvas: *mut SubCanvas,
 }
 
 impl RenderTask {
-    pub fn new(block_id: usize, tile_id: usize, subcanvas: *mut SubCanvas) -> Self {
+    pub fn new(block_id: u32, tile_id: u32, subcanvas: *mut SubCanvas) -> Self {
         Self {
             block_id,
             tile_id,
@@ -29,11 +29,11 @@ pub enum ToWorkerMsg {
 
 // todo split color and transmit it at lower priority
 pub struct SubRenderResult {
-    pub tile_id: usize,
+    pub tile_id: u32,
 }
 
 impl SubRenderResult {
-    pub fn new(tile_id: usize) -> Self {
+    pub fn new(tile_id: u32) -> Self {
         Self { tile_id }
     }
 }
